Teleport and Camera
Made a hologram of the player appear on the position the player wants to teleport to. Blue is the position is allowed and red if it's not. (Still need to implement condition for teleport not allowed)

Use Feedback Loop instead of Position Lock, on the camera, to make camera transitions smoother.